# Speedup the WebGL tests by 2.5x