--- a/src/kernels/webgl/gpgpu_context_test.ts
+++ b/src/kernels/webgl/gpgpu_context_test.ts
@@ -126,6 +126,7 @@ describeWithFlags(
         const tBeforeClear =
             gpgpu.downloadFloat32MatrixFromOutputTexture(texture, 1, 1);
         expectNumbersClose(tBeforeClear[0], 10);
+        gpgpu.gl.clearColor(1, 0, 0, 0);
         const tAfterClear =
             gpgpu.downloadFloat32MatrixFromOutputTexture(texture, 1, 1);
         expectNumbersClose(tAfterClear[0], 10);