Speedup the WebGL tests by 2.5x.

On my MacBook Pro 15inch, the WebGL 2.0 tests used to take 55 sec. Now they make 22sec. Same for WebGL 1.0.

This PR's travis run is 5 mins 53 sec [link](https://travis-ci.org/tensorflow/tfjs-core/builds/506987446). Before it was 8mins and 33 secs [link](https://travis-ci.org/tensorflow/tfjs-core/builds/506838185)

**Details**
- Cache the WebGL binaries globally instead of re-compiling after every describe.
- Fix a bug with the webgl debug mode being global, which caused most of the unit tests to run with gl.checkError() after every GL call.

DEV

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs-core/1625)
<!-- Reviewable:end -->